+++ CHANGES.md
@@ -5,7 +5,7 @@ Here you can see the full list of changes between each Pykit release.
 
 Version 0.3.8
 -----------
-- changed: uploader settings moved to `uploadSettings` config
+- changed: uploader settings moved to `uploadOptions` config
 
 Version 0.3.7
 -----------